docs: binaryService 구현
